# Conflicts:
#	console/android/GenericApp/app/build.gradle
#	console/android/GenericApp/app/src/main/AndroidManifest.xml
#	console/android/GenericApp/app/src/main/java/io/openremote/app/ui/ProjectActivity.kt
#	console/android/GenericApp/app/src/main/java/io/openremote/app/ui/SplashActivity.kt
#	console/android/GenericApp/app/src/main/res/layout/activity_project.xml
#	console/android/ORLib/build.gradle
#	console/android/ORLib/src/main/java/io/openremote/orlib/ORConstants.kt
#	console/android/ORLib/src/main/java/io/openremote/orlib/ui/OrMainActivity.kt
#	console/iOS/ORLib/ORLib.xcodeproj/project.pbxproj

+++ .ci_cd/deploy.sh
@@ -99,33 +99,18 @@ if [ -n "$SSH_USER" ]; then
   hostStr="${SSH_USER}@$hostStr"
 fi
 
-# Cannot ping from github runners so commenting this out
-# Check host is reachable (ping must be enabled)
-#if [ "$SKIP_HOST_PING" != 'true' ]; then
-#  echo "Attempting to ping host"
-#  ping -c1 -W1 -q $OR_HOSTNAME &>/dev/null
-#  if [ $? -ne 0 ]; then
-#    echo "Host is not reachable by PING"
-#    if [ "$SKIP_AWS_EC2_START" != 'true' ] && [ "$AWS_ENABLED" == 'true' ]; then
-#      "temp/aws/start_stop_host.sh" "START" "$OR_HOSTNAME"
-#      if [ $? -ne 0 ]; then
-#        # Don't exit as it might just not be reachable by PING we'll fail later on
-#        echo "EC2 instance start failed"
-#      else
-#        echo "EC2 instance start succeeded"
-#      fi
-#    fi
-#  fi
-#fi
-
 # Grant SSH access to this runner's public IP on AWS
 if [ "$SKIP_SSH_WHITELIST" != 'true' ]; then
+
+  source temp/aws/login.sh
+
   if [ -n "$CIDR" ]; then
-    if [ -z "$ACCOUNT_NAME" ] && [ -z "$ACCOUNT_ID" ]; then
+    if [ -z "$AWS_ACCOUNT_NAME" ] && [ -z "$AWS_ACCOUNT_ID" ]; then
+
       echo "Account ID or name is not set so searching for it"
       source temp/aws/get_account_id_from_host.sh
 
-      if [ -z "$ACCOUNT_ID" ]; then
+      if [ -z "$AWS_ACCOUNT_ID" ]; then
         echo "Unable to determine account for host '$HOST'"
         exit 1
       fi
@@ -134,7 +119,7 @@ if [ "$SKIP_SSH_WHITELIST" != 'true' ]; then
     source temp/aws/set_github-da_account_arn.sh
 
     echo "Attempting to add runner to AWS SSH whitelist"
-    "temp/aws/ssh_whitelist.sh" "$CIDR" "github-da"
+    "temp/aws/ssh_whitelist.sh" "$CIDR" "github-runner" "github-da"
     if [ $? -eq 0 ]; then
       SSH_GRANTED=true
     fi
@@ -143,7 +128,7 @@ fi
 
 # Determine host platform via ssh for deployment image building (can't export/import manifests)
 PLATFORM=$($sshCommandPrefix $hostStr -- uname -m)
-if [ "$?" != 0 -o -z "$PLATFORM" ]; then
+if [ $? -ne 0 ] || [ -z "$PLATFORM" ]; then
   echo "Failed to determine host platform, most likely SSH credentials and/or settings are invalid"
   revoke_ssh
   exit 1
@@ -210,7 +195,7 @@ if [ "$ROLLBACK_ON_ERROR" == 'true' ]; then
   rm -fr temp_old
   mv temp temp_old
   # Tag existing manager image with previous tag (current tag might not be available in docker hub anymore or it could have been overwritten)
-  docker tag `docker images openremote/manager -q | head -1` openremote/manager:previous
+  docker tag '`docker images openremote/manager -q | head -1`' openremote/manager:previous
 else
   echo "Removing old temp deployment dir"
   rm -fr temp
@@ -223,12 +208,6 @@ chmod +x -R temp/
 set -a
 . ./temp/env
 set +a
-
-# Login to AWS if credentials provided
-AWS_KEY=$AWS_KEY
-AWS_SECRET=$AWS_SECRET
-AWS_REGION=$AWS_REGION
-source temp/aws/login.sh
 
 if [ -f "temp/manager.tar.gz" ]; then
   echo "Loading manager docker image"
@@ -250,12 +229,15 @@ if [ \$? -ne 0 ]; then
 fi
 
 # Attempt docker compose down
-echo "Stopping existing stack"
-docker-compose -f temp/docker-compose.yml -p or down 2> /dev/null
+CONTAINER_IDS=\$(docker ps -q)
+if [ -n "\$CONTAINER_IDS" ]; then
+  echo "Stopping existing stack"
+  docker-compose -f temp/docker-compose.yml -p or down 2> /dev/null
 
-if [ \$? -ne 0 ]; then
-  echo "Deployment failed to stop the existing stack"
-  exit 1
+  if [ \$? -ne 0 ]; then
+    echo "Deployment failed to stop the existing stack"
+    exit 1
+  fi
 fi
 
 # Run host init
@@ -477,8 +459,15 @@ fi
 
 echo "Testing manager web server https://$OR_HOSTNAME..."
 response=$(curl --output /dev/null --silent --head --write-out "%{http_code}" https://$OR_HOSTNAME/manager/)
+count=0
+while [[ $response -ne 200 ]] && [ $count -lt 12 ]; do
+  echo "https://$OR_HOSTNAME/manager/ RESPONSE CODE: $response...Sleeping 5 seconds"
+  sleep 5
+  response=$(curl --output /dev/null --silent --head --write-out "%{http_code}" https://$OR_HOSTNAME/manager/)
+  count=$((count+1))
+done
+
 if [ $response -ne 200 ]; then
-  echo "Response code = $response"
   revoke_ssh
   exit 1
 fi
